add TLSUV_HTTP option to control capabilities and dependencies
